Update Unity.md
I've added line breaks and made the line an inline link. Markdown does not allow for indentation, so this poem is not perfectly reproduced.  If you really wanted it perfect, you'd use a block quote or do it in html. Please merge and resubmit the assignment for a better score.